+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 22.07.2018
 
+* [Rust] Code of Kutulu - Move dist const (CodinGame)
 * [Rust] Code of Kutulu - Improved run away algorithm (CodinGame)
 
 #### 21.07.2018